+++ src/components/UI/Keypad/Keypad/Keypad.tsx
@@ -1,31 +1,30 @@
-import Key, {KeyProps} from '../Key/Key'
+import Key from '../Key/Key'
 import styled from 'styled-components';
-import {ReactNode} from "react";
+import { ReactNode, Fragment } from "react";
 import {FiDelete} from "react-icons/all";
+import {FormattedMessage} from "react-intl";
 
 type KeypadClickHandlers = {
     numberKeyOnClick: (num: number) => void,
     deleteKeyOnClick: () => void,
-    cancelKeyOnClick: () => void,
-    unlockKeyOnClick: () => void
+    cancelKeyOnClick?: () => void,
+    unlockKeyOnClick?: () => void
+}
+
+type KeypadProps = {
+    clickActionHandlers: KeypadClickHandlers
 }
 
 const StyledKeypad = styled.div`
   display: grid;
   grid-template-columns: repeat(3, 5rem);
   grid-template-rows: repeat(4, 5rem);
-  grid-gap: 10px;
-  width: 100%;
-  height: 100%;
+  grid-gap: 20px;
 `
 
-const alphanumericKeys = (keypadClickHandlers: KeypadClickHandlers): Array<ReactNode> => {
+const alphanumericKeys = ({cancelKeyOnClick, deleteKeyOnClick, numberKeyOnClick, unlockKeyOnClick}: KeypadClickHandlers): Array<ReactNode> => {
     const alphanumericKeys: Array<ReactNode> = [];
     const alphabets = ['ABC', 'DEF', 'GHI', 'JKL', 'MNO', 'PQRS', 'TUV', 'WXYZ']
-
-    const numberKeyOnClick = (number: number) => {
-        return keypadClickHandlers.numberKeyOnClick(number)
-    }
 
     for(let i = 1; i <= 15; i++) {
 
@@ -41,50 +40,38 @@ const alphanumericKeys = (keypadClickHandlers: KeypadClickHandlers): Array<React
                 key = <Key onClick={() => numberKeyOnClick(0)} number={0} alphabet="+"/>
                 break;
             case i === 12:
-                key = <Key onClick={() => keypadClickHandlers.deleteKeyOnClick} component={<FiDelete size={24}/>}/>
+                key = <Key onClick={() => deleteKeyOnClick()} component={<FiDelete size={24}/>}/>
                 break;
             case i === 13:
-                key = <Key onClick={() => keypadClickHandlers.cancelKeyOnClick} text="Cancel"/>
+                if (cancelKeyOnClick) {
+                    key = <Key onClick={() => cancelKeyOnClick()} component={<FormattedMessage id="keypad.cancel"/>}/>
+                }
                 break;
             case i === 15:
-                key = <Key onClick={() => keypadClickHandlers.unlockKeyOnClick} text="Unlock"/>
+                if (unlockKeyOnClick) {
+                    key = <Key onClick={() => unlockKeyOnClick()} component={<FormattedMessage id="keypad.unlock"/>}/>
+                }
                 break;
             default:
                 key = <Key onClick={() => numberKeyOnClick(i)} number={i} alphabet={alphabets.shift()}/>
                 break;
         }
 
-
         alphanumericKeys.push(
             key
         )
-
-        // if (i === 1) {
-        //     key = <Key number={i} onClick={() => numberKeyOnClick(i)}/>
-        // } else {
-        //     key = <Key number={i} alphabet={alphabets.shift()} onClick={() => numberKeyOnClick(i)}/>
-        // }
     }
-
-    // alphanumericKeys.push(<Key onClick={() => numberKeyOnClick(0)} number={0} alphabet="+"/>)
-    // alphanumericKeys.push(<Key onClick={keypadClickHandlers.deleteKeyOnClick} component={<FiDelete size={24}/>}/>)
-    //
-    // alphanumericKeys.push(<Key onClick={keypadClickHandlers.cancelKeyOnClick} text="Cancel"/>)
-
     return alphanumericKeys
 }
 
-const Keypad = () => {
 
-    const keypadClickHandlers: KeypadClickHandlers = {
-        numberKeyOnClick: () => {},
-        deleteKeyOnClick: () => {},
-        cancelKeyOnClick: () => {},
-        unlockKeyOnClick: () => {}
-    }
+
+const Keypad = (props: KeypadProps) => {
     return (
         <StyledKeypad>
-            {alphanumericKeys(keypadClickHandlers).map(key => key === null ? <Key onClick={() => {}} isDisabled={true}/> : key)}
+            {alphanumericKeys(props.clickActionHandlers).map((key, idx) => key === null ?
+                <Fragment key={idx}><Key onClick={() => {}} isDisabled={true}/></Fragment> :
+                <Fragment key={idx}>{key}</Fragment>)}
         </StyledKeypad>
     )
 }