+++ src/localization/messages/en-US.ts
@@ -1,7 +1,10 @@
 import { Messages } from '../types'
 
 const en_US: Messages = {
-    'main.greeting': 'Hello world!'
+    'main.greeting': 'Hello world!',
+    'unlock.title': 'Enter your Seguro passcode',
+    'keypad.cancel': 'Cancel',
+    'keypad.unlock': 'Unlock'
 }
 
 export default en_US