+++ src/localization/messages/zh-CN.ts
@@ -1,7 +1,10 @@
 import { Messages } from '../types'
 
 const zh_CN: Messages = {
-    'main.greeting': '你好世界!'
+    'main.greeting': '你好世界!',
+    'unlock.title': '輸入您的 Seguro 密碼',
+    'keypad.cancel': '取消',
+    'keypad.unlock': '解鎖'
 }
 
 export default zh_CN