+++ src/localization/types.ts
@@ -4,7 +4,10 @@ export type Locale = (
 )
 
 export type Messages = {
-    'main.greeting': string
+    'main.greeting': string,
+    'unlock.title': string,
+    'keypad.cancel': string,
+    'keypad.unlock': string
 }
 
 export type MessagesByLocale = Record<Locale, Messages>